==> Scripts_and_Notebooks/01_Run_ECLAIR.ipynb <==
# Set the process name to be human readable in htop
import setproctitle
setproctitle.setproctitle("01_Run_ECLAIR")

from config import *

This notebook contains the command that WOULD run the ECLAIR pipeline on the provided set of interactions. The full ECLIAR pipeline is not currently provided as a standalone piece of software. For demonstrative purposes, the raw ECLAIR prediciton outputs for a handful of interactions are provided (so that the rest of the pipeline is here is runnable in some form).

- Inputs:
  - Interactions.txt


- Outputs:
  - [P1]\_[P2]\_[prot].pkl (Raw Eclair Prediction outputs)
  - Updates several feature files some of which are used downstream and are provided as static copies. Specifically...
    - uniprot_info.txt
    - pfam_domains.txt


- Dependencies:
  - Should be run AFTER all interacitons have been fed through ECLAIR pipeline for interface prediction
    - **NOTE:** The ECLAIR pipeline is not incluuded in this repository, and treats any output from this pipeline as a static result that is already available

# Theoretically run this in the command line or below using Jupyter Magic (pipeline not provided, not currently re-runnable)
print "eclair -f {0}/Interactions.txt -nc 10 -predpath {1}".format(input_dir, raw_eclair_dir)

!eclair -f /home/sdw95/3D_SARS2/git_hub/3D_SARS2/inputs/Interactions.txt -nc 10 -predpath /home/sdw95/3D_SARS2/git_hub/3D_SARS2/static_resources/Precomputed_ECLAIR_Predictions
