==> Scripts_and_Notebooks/02_Fetch_Eclair_Preds.ipynb <==
# Set the process name to be human readable in htop
import setproctitle
setproctitle.setproctitle("02_Fetch_Eclair_Preds")

from config import *

from tqdm import tqdm, tqdm_notebook
from tqdm._tqdm_notebook import tqdm_notebook

tqdm.pandas(tqdm_notebook)
tqdm_notebook.pandas()

import pandas as pd
import glob

This notebook reads in all raw outputs from the ECLAIR pipeline, parses them, and resaves them as plain text files. Additionally generates the initial data for the Interface_Summary.txt.

Raw ECLAIR outputs are saved as .pkl objects and separated out as [P1]\_[P2]\_0.pkl and [P1]\_[P2]\_1.pkl (for the interface predictions on the P1-P2 interaction for P1 and P2 respectively). The parsing done here merges the two files into one and additionally adds a confidence tier interpretation to the raw prediciton values (only High and Very High) predictions are
retained in the binary interface prediction output.

- Inputs:
  - [P1]\_[P2]\_[prot].pkl (Raw Eclair Prediction outputs)


- Outputs:
  - [P1]\_[P2].txt (Parsed Eclair Prediciton outputs)
  - Interface_Summary.txt


- Dependencies:
  - Should be run AFTER all interacitons have been fed through ECLAIR pipeline for interface prediction (01_Run_ECLIAR.ipynb)
    - **NOTE:** The ECLAIR pipeline is not incluuded in this repository, and treats any output from this pipeline as a static result that is already available

# Collect / Parse all Raw Eclair Predictions

# Store ires summary
summary = []

# Iterate over all predictions in the input directory
for f in tqdm_notebook(glob.glob(raw_eclair_dir + "/*.pkl")):
    # Read ECLAIR Predictions
    preds = pd.read_pickle(f)
    
    # Label by confidence tier
    def label_tier(p):
        if(p < 0.12):
            return "Very Low"
        elif(p < 0.24):
            return "Low"
        elif(p < 0.36):
            return "Medium"
        elif(p < 0.48):
            return "High"
        else:
            return "Very High"
    # FUNCTION END
    preds["Tier"] = preds["Pred"].map(label_tier)
    
    # Reorder so COVID protein is P1 (for consistency with other naming)
    p1, p2 = preds[["P1", "P2"]].values[0]
    if(not "COVID" in p1):
        preds["P1"] = p2
        preds["P2"] = p1
        p1, p2 = p2, p1
        preds["Prot"] = 1 - preds["Prot"]
        preds = preds.sort_values(["Prot", "Pos"])
    
    # Save Preds to dest folder
    preds.to_csv("{0}/{1}_{2}.txt".format(eclair_dir, p1, p2), sep="\t", index=None)
    
    # Store High / Very High Ires for Compiled Summary
    p1_ires = sorted(preds[(preds["Pred"] >= 0.36)&(preds["Prot"] == 0)]["Pos"].to_list())
    p2_ires = sorted(preds[(preds["Pred"] >= 0.36)&(preds["Prot"] == 1)]["Pos"].to_list())
    
    summary.append([p1, p2, "ECLAIR", sum(preds["Prot"] == 0), len(p1_ires), ",".join([str(x) for x in p1_ires]), sum(preds["Prot"] == 1), len(p2_ires), ",".join([str(x) for x in p2_ires])])

# Create final DF
summary = pd.DataFrame(summary, columns=["P1", "P2", "Source", "P1_Len", "P1_N_Ires", "P1_Ires", "P2_Len", "P2_N_Ires", "P2_Ires"])

summary.head()

# Save summary file
summary.to_csv("{0}/Interface_Summary.txt".format(output_dir), sep="\t", index=None)
